Add unsubscribe instructions to notification email
Note that I would simply link to `https://giscus.co/user/confirm-cancel`. However, if not logged in, that redirects to `/login` which currently returns a 404. The only way to log in is apparently to first visit the home page and click the GitHub auth button. From there, a user can proceed to the cancellation confirmation page.

Closes #31